docs

+++ tests/pyi/fixtures/general/compile_time_shape_exprs.pyi
@@ -4,5 +4,5 @@ n1: Final[Int32]
 
 def use_expr(
     x: Int32[n1 - 1 - 0 + 1],
-    y: Float64[n0 * 2]
+    y: Float32[n0 * 2]
 ) -> None: ...
